頑張った

+++ EnemyData.h
@@ -10,7 +10,7 @@ struct EnemyData {
 	int SPD;
 	int Gold;
 	int EXP;
-	int CriticalRate;
+    float CriticalRate;
 	Element Element;
 
 };
